requirements.txt
fastapi[all]
uvicorn
sqlalchemy
alembic
asyncpg
fastapi-cache2[redis]
celery
flower
pytest-asyncio
GeoAlchemy2
greenlet
argon2-cffi
psycopg2-binary
python-dotenv
python-jose[cryptography]
aiosmtplib
email-validator
boto3
aiobotocore
openai
langchain
loguru
tenacity


